iommu/samsung: add support for iova reserve
iova areas need to be reserved for some reasons:
We can imagine a DMA device that is running since bootloader and needs
to keep running after kernel booting. It was running in the physical
address space before kernel booting and should seamlessly run after its
IOMMU is attached. For those devices, the DMA area the device is
accessing should be reserved from the bootloader and the iova area with
the identical address of the DMA area should be written in the page
table of the IOMMU when it is attached.
Another reason for the reserved iova area is for a DMA device that needs
some buffers should be in a specific area. One of the example is MFC.
The firmware address should be lower than all other buffers MFC should
access. To guarantee the restriction, MFC driver should reserve the DMA
address for the firmware before DMA addresses are allocated for other
buffers.

The "samsung,iommu-identity-map" property reserves and maps the
specified DMA area for the former reason while
"samsung,iommu-reserved-map" just reserves the specific area for the
latter reason but does not map it.

These properties should be specified in the device tree node of the
client devices. The base address of a region should be #address-cells
wide and the size of the region should be #size-cells wide.

Here is an example of a device that has two reserved identity mappings.

client {
    ...<snip>...
	    samsung,iommu-identity-map = <0x0 0x10680000 0x10000>,
					 <0x0 0x10700000 0x10000>;
};

Change-Id: I73a398f2f8fad460c19db93abbf43488b09ce711
Signed-off-by: Janghyuck Kim <janghyuck.kim@samsung.com>
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>
Signed-off-by: Jinkyu Yang <jinkyu1.yang@samsung.com>
Signed-off-by: Daniel Mentz <danielmentz@google.com>
(cherry picked from commit fd7b428e357d2f85c4a21e2acdeb62c7cca277c6)
Signed-off-by: oscyre1006627 <165604309+oscyre1006627@users.noreply.github.com>

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>