media: mfc: cache flush is needed for MOVE_INSTANCE
CACHE_FLUSH is required for the MOVE_INSTANCE command,
but the cache_flush() was not working because the cache_flush()
was synchronized according to the DRM attribute conversion.
When only the DRM attribute is changed, it to check
if the previous F/W is loaded or not.

Bug: 234321602
Bug: 238360304
Test: Build PASS
Signed-off-by: AyoungSim <a.sim@samsung.com>
Signed-off-by: wenchangliu <wenchangliu@google.com>
Change-Id: Ia52f713cb25b091242426b93fcf0316af5d0be9b
(cherry picked from commit 3ef5f5737d0debf90bdd3318ac22be3c739b775a)

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>